Rename Map component to LeafletMap
