+++ components/LeafletMap/index.ts
@@ -1,3 +1,3 @@
-export * from './Map';
+export * from './LeafletMap';
 export * from './TileLayer';
 export * from './GeoJSONLayer';
